+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## 7.3.0 [unreleased]
 
+
 ### Bug Fixes
 
 - [#807](https://github.com/influxdata/influxdb-client-java/pull/807): fix: Issue/806: Add a lock to prevent the disposal of the timer while it is running.
+
+
+### Features
+
+- [#821](https://github.com/influxdata/influxdb-client-java/pull/821): Prevent duplicate interceptors in OkHttpClient builder
 
 
 ### Dependencies